Support pre-creating Histogram object so that it may be reused when repeatedly processing images.
This object is 47,436,840 bytes in size, so will be allocated on the LOH. Running many quantizations without this patch results in lots of garbage in the LOH and, in my experience, OutOfMemoryExceptions.

With this patch, the caller creates an instance of Histogram then passes it for each quantization operation. In this way, memory usage is flat. This maintains the thread safety of the WuQuantizer class too, as data is kept isolated between threads.

This is a non-breaking change to the API. A new overload is added.