fix: Value of '-3' is not valid for 'alpha'.
When dealing with large files, there is possibility that if there are too many pixels with the same colour, it will wrap the int.
This is particularly likely with the Alpha value because a non-transparent image will have values of 255 for each pixel.

This changed the variables in the PaletteColorHistory to be ulong which also mitigates the error by not allowing negative values (if it wraps, it would be > 0 still).